+++ Exp-4/maven-app/demo-app/src/main/java/com/example/App.java
@@ -1,0 +1,13 @@
+package com.example;
+
+/**
+ * Hello world!
+ *
+ */
+public class App 
+{
+    public static void main( String[] args )
+    {
+        System.out.println( "Hello World!" );
+    }
+}
