--- Exp-4/maven-app/src/main/java/com/example/App.java
@@ -1,13 +1,0 @@
-package com.example;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
